Put prop_conv_solvert into separate file
Separates the interface and (one of) its implementations.
